Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
requests
python-multipart
passlib[bcrypt]
python-jose[cryptography]
faiss-cpu
sentence-transformers
python-pptx
